+++ NEWS.md
@@ -6,7 +6,7 @@
 ### Enhancements
 * Added `ADQLQC`, `ADCSSRS`, and `ADEQ5D5L` as supported data sets.
 * Improved error message in `get_cdisc_keys`.
-* Examples now use `scda.2022` instead of `scda.2021`.
+* Removed `scda` package dependency from examples.
 * Fixed help files for `TealDataset` and `MAETealDataset`.
 * Added backstop for missing `reticulate` package in _teal.data with Python_ vignette.